+++ common/src/main/java/dev/boxadactle/coordinatesdisplay/command/CoordinatesCommand.java
@@ -1,71 +1,42 @@
 package dev.boxadactle.coordinatesdisplay.command;
 
-import com.google.common.collect.ImmutableList;
 import com.mojang.brigadier.context.CommandContext;
 import dev.boxadactle.boxlib.command.BCommandSourceStack;
-import dev.boxadactle.boxlib.command.api.BClientCommand;
+import dev.boxadactle.boxlib.command.api.BCommand;
+import dev.boxadactle.boxlib.command.api.subcommand.BasicSubcommand;
 import dev.boxadactle.boxlib.scheduling.Scheduling;
 import dev.boxadactle.boxlib.util.ClientUtils;
-import dev.boxadactle.boxlib.util.GuiUtils;
 import dev.boxadactle.boxlib.util.WorldUtils;
 import dev.boxadactle.coordinatesdisplay.CoordinatesDisplay;
+import dev.boxadactle.coordinatesdisplay.position.Position;
 import dev.boxadactle.coordinatesdisplay.screen.CoordinatesScreen;
 import dev.boxadactle.coordinatesdisplay.screen.config.PositionScreen;
-import dev.boxadactle.coordinatesdisplay.position.Position;
-import net.minecraft.network.chat.Component;
-
-import java.util.List;
+import net.minecraft.client.resources.language.I18n;
 
 public class CoordinatesCommand {
 
-    public static BClientCommand create() {
-        return BClientCommand.create("coordinates", CoordinatesCommand::openCoordinatesScreen)
-                .registerSubcommand(new ConfigSubcommand())
-                .registerSubcommand(new CornerSubcommand())
-                .registerSubcommand("help", CoordinatesCommand::showHelpMessage)
-                .registerSubcommand(new ModeSubcommand())
-                .registerSubcommand(new VisibilitySubcommand())
-                .registerSubcommand("movehud", CoordinatesCommand::moveHud)
-                .registerSubcommand(new PositionSubcommand())
-                .registerSubcommand("toggle", CoordinatesCommand::toggle);
+    public static BCommand createCommand() {
+        return BCommand.create("coordinates", CoordinatesCommand::openCoordinatesScreen)
+                .registerSubcommand(ToggleSubcommand.create())
+                .registerSubcommand(MoveHudSubcommand.create())
+                .registerSubcommand(ConfigSubcommand.create())
+                .registerSubcommand(CornerSubcommand.create())
+                .registerSubcommand(ModeSubcommand.create())
+                .registerSubcommand(VisibilitySubcommand.create())
+                .registerSubcommand(PositionSubcommand.create())
+                .registerSubcommand(TeleportModeSubcommand.create());
     }
 
-    private static int openCoordinatesScreen(CommandContext<BCommandSourceStack> ignored) {
+    static int openCoordinatesScreen(CommandContext<BCommandSourceStack> ignored) {
         Scheduling.nextTick(() -> ClientUtils.setScreen(new CoordinatesScreen(Position.of(WorldUtils.getPlayer()))));
 
         return 0;
     }
 
-    private static int toggle(CommandContext<BCommandSourceStack> ignored) {
-        CoordinatesDisplay.getConfig().enabled = !CoordinatesDisplay.getConfig().enabled;
-        CoordinatesDisplay.CONFIG.save();
+    static int noArgs(CommandContext<BCommandSourceStack> ignored) {
+        CoordinatesDisplay.LOGGER.player.error(I18n.get("command.coordinatesdisplay.emptyArgs"));
 
-        return 0;
-    }
-
-    private static int showHelpMessage(CommandContext<BCommandSourceStack> ignored) {
-        List<Component> components = ImmutableList.of(
-                GuiUtils.colorize(Component.translatable("command.coordinatesdisplay.helpmenu"), GuiUtils.AQUA),
-                Component.translatable("command.coordinatesdisplay.config"),
-                Component.translatable("command.coordinatesdisplay.gui"),
-                Component.translatable("command.coordinatesdisplay.help"),
-                Component.translatable("command.coordinatesdisplay.mode"),
-                Component.translatable("command.coordinatesdisplay.movehud"),
-                Component.translatable("command.coordinatesdisplay.position"),
-                Component.translatable("command.coordinatesdisplay.visibility")
-        );
-
-        components.forEach(c -> {
-            CoordinatesDisplay.LOGGER.player.chat(c);
-        });
-
-        return 0;
-    }
-
-    private static int moveHud(CommandContext<BCommandSourceStack> ignored) {
-        Scheduling.nextTick(() -> ClientUtils.setScreen(new PositionScreen(null)));
-
-        return 0;
+        return 1;
     }
 
 }