+++ fabric/src/main/java/dev/boxadactle/coordinatesdisplay/fabric/CoordinatesDisplayFabric.java
@@ -11,6 +11,7 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
+import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.world.entity.player.Player;
@@ -43,7 +44,7 @@ public class CoordinatesDisplayFabric implements ClientModInitializer {
         }
     }
 
-    private void renderHud(GuiGraphics guiGraphics, float f) {
+    private void renderHud(GuiGraphics guiGraphics, DeltaTracker deltaTracker) {
         try {
             if (CoordinatesDisplay.HUD.shouldRender(CoordinatesDisplay.getConfig().visibilityFilter)) {
                 ModConfig config = CoordinatesDisplay.getConfig();