+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## Changes
-- Backport to 1.20.6
-- Fixed hotbar renderer in config screens
+- Port to 1.21
+- Removed help subcommand
+- Added relative boolean to corner subcommands
+- Opening config file with command now errors to chat
+- Added teleport mode subcommand
+- Visibility, mode, and corner subcommands log to chat
+- Integers can now be specified in movehud subcommand
+- Boolean can now be specified in toggle subcommand